clean up some bugs.
1. changes to elifeBinaryRandomForestResults and plotROC.py
- The data load function us to return X as numpy array, it is now a data frame.
Using a dataframe instead of an array causes the model to save not only
the feature names but their expected order. This will reduce a lot of
bugs in the future

- when we upgrade our version of python. It broke scikit-plot.
environment-scikit-plotHack.yml has a hack work around. basically it uses
an old version of scipy. TODO https://github.com/mljar/mljar-scikit-plot

TODO: double check labelEncoder. we want healthy to be zero and diease to be 1